[451365] Adds specific accessors for OpaqueElementBodyChanges
For moves of language/body values, we show the list of languages in the
order of each side. For additions, changes, and deletions, we show the
added, deleted, or changed body value using a textual content merge
viewer.

Therefore, we add two separate accessors
(OpaqueElementBodyChangeAccessor and OpaqueElementBodyMoveAccessor),
which are created by the OpaqueElementBodyChangeAccessorFactory
depending on the opaque element body change's difference kind.

Bug: 451365
Change-Id: I05aa142475357896e0303687171babab835ee719
Signed-off-by: Philip Langer <planger@eclipsesource.com>