[454579] Fix Nullpointer when element is moved out of FeatureMap
The FeatureMapChangeMerger is intended to handle manipulations of
FeatureMaps. When used to work on the Non-Feature-Map side of an
operation a NullPointerException is thrown. This patch changes the
AbstractMerger so that equivalent ReferenceChanges will be preferred in
this scenario.

The AbstractMerger is also modified to first search for the master
equivalence and only after that to check for equivalences with
conflicts. This avoids an issue where multiple equivalences with
conflicts exist and the less fitting one is chosen.

Includes 2-way and 3-way testcases.

Bug: 454579
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I8a264b09a7a8011121c341673638eda052afb422