[451365] Removes conflicts and implications of refining diffs
The AbstractChangeFactory propagates conflicts from refining diffs to
refined diffs. For OpaqueElementBodyChanges there is a specific conflict
detection, so conflicts of refining changes are not relevant.
Especially, since the typcial refining diffs often have pseudo
conflicts. If these pseudo conflicts are then propagated to the
OpaqueElementBodyChanges, they will be filtered in the structured diff
viewer.

To address this issue, we now remove all propagated conflicts from
OpaqueElementBodyChanges.

Also, we used the implies relationship to make sure that the refining
changes are handled when the OpaqueElementBodyChanges is merged.
However, this had implications on the user interface of the structure
merge viewer.

To address this issue, we removed those implications but now we have to
remove the refining diffs from the merge dependencies; otherwise, when
the opaque element body change is merged, AbstractMerger would first
invoke the merge of the refining diffs as they are considered as merge
dependencies, which causes the merge of the entire
OpaqueElementBodyChange anyway, and then pursue the merge of the
OpaqueElementBodyChange, so that it would ultimately be performed twice.
Therefore, we remove the refining diffs from the merge dependencies by
overriding #getDirectmergeDependencies(Diff, boolean).

Bug: 451365
Change-Id: Ib61e2564665f83c9d076293706df63e024b29e63
Signed-off-by: Philip Langer <planger@eclipsesource.com>