[453218] Add missing FeatureMapChanges concerning Moves
Fixes the DefaultDiffEngine to also add FeatureMapChanges of
DifferenceKind.MOVE where the corresponding element resides in a
Non-FeatureMap-Containment in the left-hand-side while residing in a
FeatureMap-Containment in the right-hand-side. 

Also modified are the DefaultEquiEngine to properly detect equivalences
concerning the new diffs and the FeatureMapChangeMerger to correctly
merge the new diffs.

Includes testcases.

Bug: 453218
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I1a9721ce3265d43f2b97a0f992ff6915731fbcf1