[456860] Improve handling of equivalent three-way text changes
Three-way diffing and patching is not supported by google diff match
patch (dmp) and simply applying the patch computed by dmp to the
opposite side, as we did before, would cause the equivalent changes to
be applied twice. Therefore, we had to re-implement the detection and
merging ourselves to support the handling of equivalent changes.

We also adapted the test cases and the implementation of the three-way
diff and merge to correspond to the behavior of the org.eclipse.compare
line differencing and introduced a determination of the currently used
line separator in the text, which is now also used when building the
merged version.

Bug: 456860
Change-Id: I12be9c68eda663e0206c9f230a9faf9631be3275
Signed-off-by: Philip Langer <planger@eclipsesource.com>