[461986] Intermediate Proxies Resolution
When a model is split into fragments, and only some of them have
changes, the structure merge viewer (SMV) and the content merge viewers
(CMV) display the models involved in the comparison but don’t display
the fragments that have no changes.

If a change (x) is detected in a fragment (B), and this fragment is a
child of another fragment (A) that has no changes, then (A) won't appear
in the SMV and the CMV's. As a result, users will think (B) is the root
of the global model.

To avoid this, the idea is to display intermediate node(s) (represented
as [...]) in order to show to users that it exists something (fragments,
i.e. a parts of models) between/above the changes.

Bug: 461986
Change-Id: I8ca496cbed573cc48c0f163816beb951d073165f
Signed-off-by: Axel Richard <axel.richard@obeo.fr>