Do not remove read-only storages from comparison scope
It is possible for a resource to have been added on one of
the remote sides. In such a case, the resource is both
read-only (remote side) and out of the workspace (it doesn't
even exist locally). Still, there might be "move" difference
that we need to detect, and we need these resources to
detect them properly (see associated bug).

Bug: 460773
Change-Id: I728a97353e884d63e117f3a8fa4479db99baeb9c
Signed-off-by: Laurent Goubet <laurent.goubet@obeo.fr>