Refactoring of ThreadedModelResolver.
Split computations in several smaller classes to ease maintenance.
Extracted multi-threaded logic in a separate class.
Extracted function AS_URI to ResourceUtil.
Extracted IResourceDependencyProviderListener in its own class
Factor out Ecore and URI dependency from DependencyFoundEvent
Encapsulates decision on whether a parent URI is registered (i.e.,
whether the reference is containment or not) in a new subclass
ResourceDependencyFoundEvent. This also allows to remove dependencies to
URI and EcoreUtils.
To allow mocking the graph, we also change Graph to not be final
anymore.
Added GraphResolutionTest to test suite.
Extracted multiple interfaces to separate:
- dependency obtention;
- local resolution;
- remote resolution.

Change-Id: I7fe11b2cf6f4abc13f64821826855fe6fe205ef2
Also-by: Philip Langer <planger@eclipsesource.com>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Axel Richard <axel.richard@obeo.fr>