[464904] Ensure NotLoadingResourceSet looks at all proxies
Replaces the proxy map of the NotLoadingResourceSet with a queue to
ensure all proxies are looked at.

Previously when resources were loaded during the resolving process they
were added to the proxy map while the NotLoadingResourceSet was
iterating over it. This lead to undeterministic behavior like sometimes
skipping all or some additionally added proxies. 


Bug: 464904
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: Ieae31537a8129fc0403f43cba8feb4a6961ecc00