[465329] Return the URI of the resource to be created if not controlled
If the source object of a resource attachement change's match has no
target match and no container, a new dependent resource has been created
at the source side. Thus, the target URI should be the original URI of
the source object; that is, the URI of the resource to be created.

Bug: 465329
Change-Id: I2f3c70bcb19f5ccd1cdae063409094f5c5e7f300
Signed-off-by: Philip Langer <planger@eclipsesource.com>