[462237] Merge Viewer: Jump to next unresolved difference after merging
After merging a difference in the merge tool, the structure viewer now
jumps to the next unresolved difference, instead of simply to the next
difference. In the process of adding said functionality, the Navigatable
class has been tidied up and refactored for better readability and
maintainability. Tests are adapted accordingly.

Bug: 462237
Change-Id: Ifb687ceebc8502805559662665d2dee0e3aeef3f
Signed-off-by: Michael Borkowski <mborkowski@eclipsesource.com>