[462884] Allows merging specific diffs in MergeAllNonConflictingRunnable
Renames MergeAllNonConflictingRunnable to MergeNonConflictingRunnable,
as it now allows to also merge a specific set of differences with the
same strategy as for all non-conflicting differences. Therefore, this
runnable now also implements IMergeRunnable.

Accordingly, also the test class for MergeNonConflictingRunnable have
been extended concerning the use cases for merging a given list of
differences only.

Also we introduce additional two-way tests in MergeAllCommandTests.

Bug: 462884
Change-Id: I80518d479db18c19bd51ebdabe5789ca58850672
Signed-off-by: Philip Langer <planger@eclipsesource.com>