Caching - Don't minimize the logical model eagerly
We use a cache to make sure the EMFModelProvider can always return the
logical model of files as fast as possible while reducing I/O operations
to the bare minimum. Minimizing the logical model _before_ we cache it
means we'll only be able to use that cache for the minimized set of
resources, forcing us to recompute the whole model for all *other*
resources.

Change-Id: I65612d9a53b454a288926a3533d57910b13a7e32