Caching - Don't minimize the logical model eagerly
We use a cache to make sure the EMFModelProvider can always return the
logical model of files as fast as possible while reducing I/O operations
to the bare minimum.
Before,we used to minimize the logical model _before_ we cached it,
which prevented minimized resources from being cached.
This change makes sure that all the resources of the logical model are
cached before the logical model is minimized.
The expected performance impact is important, our test cases show a gain
of ~10 seconds for a merge operation that took roughly 18 seconds
and takes 8 with this fix.

Change-Id: I65612d9a53b454a288926a3533d57910b13a7e32