[471010] Fix log4j config management
The log4j config initialization was made in rcp.ui which may not start
before an EGit merge operation takes place, for instance.
Moving this to rcp core solves the issue.
Additionally, the preferences that indicate the max size of log files
and the max number of log files to keep were systematically erased each
time eclipse was restarted.

Bug: 471010
Change-Id: Id05efc971853bd080425ffce7ab7b176b8744734
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>