[474723] Introduce IEMFComparisonSource to support custom comparisons
Clients can now add an adapter for 'IEMFComparisonSource" to their
objects. This adapter is used by EMFCompare to offer a "Compare with
each other" action for selections adapting to 'IEMFComparisonSource'
(but not to IResource to avoid interfering with Platform/Compare).

An adapter for IPapyrusFile is added to support PapyrusNodes when their
IResourceAdapter is eventually removed.

Bug: 474723
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I216b1b61a9f7b730bc4aecb6d06f6a2b5f6cf1f7