[475401] Modify RootIDMatchingStrategy to only return unique mappings
The RootIDMatchingStrategy determines matching between resources by
intersecting the IDs of their root objects. The RootIDMatchingStrategy
now checks if a possible mapping is unique and does therefore not
intersect with another possible mapping. 

The previous behavior of returning the first possible mapping creates
"random" results, affected by the order the given iterators return their
resources. Since this strategy can not determine which is the "better"
mapping of multiple options it should not return any of them. This way
the remaining EMFCompareFramework is not confused by "half-true"
mappings.

Includes two-way testcases.

Bug: 475401
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I2cb886d088d6ce27cbc06ec2958fc02ef9253d36