[467677] Removes dependency to jgit from ide.tests
The usage of SmallStream has been introduced to test as closely as
possible to the intended usage of ResourceUtil.binaryIdentical. However,
looking at the implementation of SmallStream, it is just a wrapper
anyway, which perfectly implements the protocol of any valid input
stream. So the usage of SmallStream is technically not necessary.

Bug: 467677
Change-Id: I1cb09a257af028305122e430f28e8f6299068f55
Signed-off-by: Philip Langer <planger@eclipsesource.com>