[475890] Add AbstractGitMergeTestCase and test case for 475890
This commit introduces AbstractGitMergeTestCase to ease creating further
integration test cases with Papyrus and Git. The idea of this abstract
test case is to be able to just add a merge scenario in the form of the
models for the origin, left, and right side, implement the validation of
the expected resulting merged model files and therewith create a test
case that covers merge, rebase, and cherry-pick in both directions.

As an example, this commit adds the test case for Bug 475890. This bug
was fixed in previous changes, but this commit still adds the test case
to avoid regression.

Bug: 475890
Change-Id: I32e7c666a33f9cfb08024080f42852bc7b421da5
Signed-off-by: Philip Langer <planger@eclipsesource.com>