[462863] Staging Prompt after Resolving Conflicts
After resolving conflicts in the (Compare) Merge Tool, the user is now
prompted with a dialog stating that all conflicts in the model have been
resolved, and asking whether the files containing the changes should be
staged.

This is not only an improvement saving some users' clicks, it also helps
to avoid staging so far still unmerged model resources. From a
user-perspective, it is not obvious which model resources are part of
the logical model being currently merged in the merge tool. So if there
is a conflicting merge that needs manual resolution, the user opens a
conflicting resource in the merge tool. After conflicts have been
resolved, users want to stage all model resources that have been
resolved currently in the merge session, but it is hard to identify
those that they can safely stage already. So users are inclined to stage
not merged models, which basically leads to omitting all changes on the
remote side.

Bug: 462863

Change-Id: Ia4f2712508b21acd3c469c35e8dfd9b3831bcca9
Signed-off-by: Michael Borkowski <mborkowski@eclipsesource.com>