[479449] Raise conflict for addition to single-valued containment ref
If there is no value in a single-valued containment reference and a new
value is set on both sides, it cannot be merged commutatively. However,
EMF Compare did not raise a conflict so far. With this change, a
conflict is raised on such changes. A test case is included.

Note that I didn't put the predicate
ConflictDetectionTest.setOfReference() in EMFComparePredicates in this
change to allow for merging this change without breaking the API. I move
it into EMFComparePredicates in a subsequent change.

Bug: 479449
Change-Id: Ic7d2ea91ca87b0fb47ac0b77db7a87eb78cce5ca
Signed-off-by: Philip Langer <planger@eclipsesource.com>