[462863] Staging Prompt after Resolving Conflicts
This change is a revised version of change 52072, which was reverted due
to undetected issues. The issues included the fact that the dialog was
shown despite the fact that no conflicts were involved, and also that
files to be staged were listed multiple times. The changes w.r.t. change
52072 were done in the following files:

* MergeResolutionManager.java (lines 76-79, an if block was added)
* PostMergeStager.java (line 58, a set is used instead of a list)

Following is the original change content:

After resolving conflicts in the (Compare) Merge Tool, the user is now
prompted with a dialog stating that all conflicts in the model have been
resolved, and asking whether the files containing the changes should be
staged.

This is not only an improvement saving some users' clicks, it also helps
to avoid staging so far still unmerged model resources. From a
user-perspective, it is not obvious which model resources are part of
the logical model being currently merged in the merge tool. So if there
is a conflicting merge that needs manual resolution, the user opens a
conflicting resource in the merge tool. After conflicts have been
resolved, users want to stage all model resources that have been
resolved currently in the merge session, but it is hard to identify
those that they can safely stage already. So users are inclined to stage
not merged models, which basically leads to omitting all changes on the
remote side.

Bug: 462863

Change-Id: Ice31ce3dde4b6f45de0f7cf8d90be00806341a55
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>
Also-by: Michael Borkowski <mborkowski@eclipsesource.com>