Remove dangerous & useless Post-Processor
When ResourceLocationChange has been abandoned, this post-processor has
been update to use ResourceAttachmentChanges instead. That is not what
should have been done: A ResourceAttachmentChange is not related to a
creation/deletion/move of a resource.

Additional tests must be performed to determine whether the complex
use-cases with UML, notation and di files still work properly.

Change-Id: I90b2a5b8b323ca27cb3c4b69b114816609709c08
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>