[487306] Use only relevant diagnostic to compute traversals
The choice of creating a traversal or a singleton traversal is made
using the diagnostic of the whole synchronisation model.
This can lead to errors since it can contain errors on resources not
directly linked to the actual comparison.
To fix this issue, a correct traversal is always created, the potential
errors are managed later by EMFCompare.

Bug: 487306
Change-Id: I982e8122515e2283800cb9799ecfa60ab8ddeac2
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>