[488551] Add hook for initializing the ElementTypeSetConfigurationReg.
The class ElementTypeSetConfigurationRegistry is not available on Luna,
so we have to use reflection to be backwards compatible with Luna.

This also fixes implementations of
o.e.papyrus.infra.core.resource.AbstractBaseModel, which now forces its
subclasses to implement canPersist(EObject) and persist(EObject). In
order to be compatible with earlier versions, we omit the @Override
annotation.

Bug: 488551
Change-Id: If5b90c3d1a27af25e4d88460686e4ef5a00194d5
Signed-off-by: Philip Langer <planger@eclipsesource.com>