[488089] Fix random position of diagrams after merge
In a three way merge, sometime when you merge resource attachment
changes of kind "move" the operation doesn't give the result you would
expect. In fact, in the resulting merged file, the order of the elements
is not respected. The order depends only in which order you merge the
change instead of using the initial position in the file.

Bug: 488089
Change-Id: I0b1ebac101f845d118cedc0341017c256d35b6ac
Also-by: Philip Langer <planger@eclipsesource.com>
Signed-off-by: Simon Delisle <simon.delisle@ericsson.com>
Signed-off-by: Philip Langer <planger@eclipsesource.com>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>