Fix and optimize usage of StorageTraversal.getStorages()
StorageTraversal.getStorages() returns a mutable copy of the internal
set of storages. This was, however, wrongly communicated its JavaDoc and
has hence also been used here and there under the assumption that this
is the actual set of storages rather than a copy leading to non-ideal
usages.

This change fixes the JavaDoc and improves the usages of this method.

Change-Id: I556a1fa6e497bdc0852c59515840d4eb9907eb23
Signed-off-by: Philip Langer <planger@eclipsesource.com>