[471045] Fix ArrayIndexOutOfBoundsException in model resolution
If we skip setValue() for particular features, then we also have to skip
setManyReference() for them. Otherwise, setManyReference() may throw an
ArrayIndexOutOfBoundsException by trying to insert values at invalid
indexes (because expected values are missing) into many-valued features.
This happenned when a non-containment EList had several elements with
some of them resolved and others not yet resolved. Then, on
endDocument(), the management of forward references invokes
setManyReference(), expecting that elements already present are indeed
present in the targeted EList, which is not the case with our parser
pool.

Bug: 471045
Change-Id: I995a5f8abc0ed32d68bb7bea70c850f78fda8beb
Signed-off-by: Andreas Mayer <anma-e@gmx.de>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>