[493650] Fix order of merged children
Improved after discussion with Laurent Goubet.
A test is added to check that when merging every diff
and resolving conflicts by:
- first reject all target side conflicting diffs
- then accepting all other diffs
yields models to be identical on both sides, no matter
in what order the diffs are accepted/rejected.
The test uses Collections2.permutations,
which appeared in guava 12, so an update of the TP was necessary.

Bug: 493650
Change-Id: Ib790b992184f6038806bf890f7af96ae2c447ca7
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>