[495334] [Testing Framework] Ancestor not loaded correctly
Adapt TestSupport class to correctly query the given ancestor parameter.
If no parameter is given, i.e., we have the empty String from the
default parameter, then no ancestor is loaded. If the parameter is
given, we attempt to load the given resource like we do for the left and
right resource. 

Bug: 495334
Change-Id: I87bd4b4ac18cdfade9dd281d6a25284fa3d86911
Signed-off-by: Martin Fleck <mfleck@eclipsesource.com>