Bug 495741 - Fix computation of visible and filtered diffs
If a node has been filtered out, then refreshTitle() now also considers
all the descendants of that node as hidden. So the number of filtered
diffs in the title should match the actual number of diffs filtered from
the viewer -- except a diff is shown in two different sub-trees and only
filtered in one of them.

Change-Id: I171552743a4d8c3d702f7830d03098755e688b4d
Signed-off-by: Andreas Mayer <anma-e@gmx.de>