Do not consider a container deletion to conflict with unsets under it
Since a containment deletion and the "unset"(s) under it were seen as a
pseudo conflict with each other, we were "hiding" the deletion from view
and preventing the merge altogether.

This adds an explicit dependency between the core tests and objenesis to
avoid failures when launching the build locally (mockito has a transient
dependency to objenesis).

Change-Id: I8d00c336285480962499d861575815f1c4b72b3d