Change SMV groups algorithm
All groups (default, bySide, byResource, byKind) are affected.
The new algorithm focuses on displaying diffs in a better way
and on avoiding as far as we can duplicates in the created groups.

This change leads to filter modifications: filters no longer have a
special behavior when deactivated, they are plain filters.

Change-Id: I6ef96eab4e6a6bf6ffb1e2ac2f3f70b79a641f59
Also-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>