Split tests to isolate dependencies to papyrus
We want to isolate the dependencies to papyrus in dedicated
projects/plug-ins so that it's easier to move them to papyrus later on.

NOTE: This commit also fixes a test that had been broken by commit
38670759ff (automatic migration of UML profiles), which prevented the
suite AllTests from being called. Only AllUITests was actually called,
because the opm declared the surefire plugin twice, which is apparently
not supported...

Change-Id: I389cfb6114a789b7aa4419321a54503dafda0d1d
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>