Fix cascading filter for special cases
To allow group extenders to properly modify the tree structure, this
filter must use the TreeNode hierarchy instead of the match hierarchy,
otherwise it will not filter the relevant elements.
It's also important to take into account the refining diffs of the diff
to filter in case they contain containment RefChanges of parent matches.

Change-Id: I431429520816073b4fd73e8e8367337d04be10c8
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>