[442897] Fix markers issue in SMV
Some markers of required or rejected diffs don't have the correct
color in the gutter. It was previously fixed by review 37012 but this
fix finally introduces other problems. This commit reverts this previous
fix and adopts the following solution: pseudo conflicts are simply not
taken into account when computing merge dependencies since there is
nothing to do with them, the changes have already been made on both
sides.

Change-Id: I554ccb6c3c6617882ba4420bfc3f0cd7f6784f8c
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>