[501864] Do not add refined diffs to conflicts of refining diffs
Refined diffs should not be added automatically to the conflicts of its
refining diffs. In case of multiple conflicts, the refined diff would
otherwise randomly end up in the last one to which it has been added.

This change also adapts the test cases and improves the
MultiplicityElementChangePostProcessor to be more efficient.

Note that I had to ignore the AdditiveMergeDiagramTests as they seem to
fail after this change. Unfortunately, I don't see why. I assume that
this is because previously the AdditiveConflictMerger handled the
refined diffs if one of its refining diffs was conflicting, but doesn't
anymore after this change.

Change-Id: Iaadf6c245cbfd69d30f981cd1934e402ecbd605a
Signed-off-by: Philip Langer <planger@eclipsesource.com>