[501864] Adapts diff filters for conflicting refined diffs
If a refined diff has a direct or indirect real conflict or all its
refining diffs have pseudo conflicts, they should not be shown in the
difference groups (left/right side). However, refined diffs that are
redefined by both diffs that have pseudo conflicts and diffs that do not
have any conflicts should be shown in the difference groups.

Bug: 501864
Change-Id: I9dd640e3736d0ca98ef194965ee8f62a3404c06c
Signed-off-by: Tanja Mayerhofer <tmayerhofer@eclipsesource.com>