[507050] Fix conflict merging when pre-merge is set to off
When conflicts occur in only some files of a logical model, differences
in non-conflicting files of the model will be lost when pre-merge is set
to off.

This is avoided by setting all files with differences to conflict when
pre-merge is set to off and they are part of a conflicting model.

Bug: 507050
Change-Id: I284a974a6c3a4991b37eca44f8cd9cd83b7b4ddc
Also-by: Philip Langer <planger@eclipsesource.com>
Also-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>