[497566] Never alter files before save
EMF Compare now installs an Adapter on the left and right ResourceSet
and Resources to react to content changes. When a Resource is added or
becomes empty, the Adapter checks whether the resource must be deleted
and keeps track of that info. It also reacts when undo/redo are
performed. Resources are actually only deleted on save.

Additionally a mechanism on the new ResourceChangeAdapter allows
third-parties to participate in the decision. That way it's possible to
create or delete several resources at a time, and to prevent a deletion
of one resource if some associated resources must not be deleted. This
helps supporting Papyrus for example.

Change-Id: Ic77e2482e4db5ec537126e3e9c0fa96d633247ce
Also-by: Axel Richard <axel.richard@obeo.fr>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>