[508665] Delete EnumerationLiteral classifier changes from comparison
EnumerationLiteral.classifier is essentially a derived feature. In fact,
the classifier of an EnumerationLiteral is always its containing
Enumeration. However, this feature is not declared as such in the
metamodel, as it is inherited from InstanceSpecification.classifier,
which is not derived. In the UML implementation,
EnumerationLiteral.classifier, therefore, overwrites
InstanceSpecification.classifier to return an unmodifiable collection
always containing the Enumeration instead of a writable collection.

Thus, we have to avoid merging changes of EnumerationLiteral.classifier
changes and can ignore changes of this feature altogether. Changes of
this feature on EnumerationLiterals will therefore be deleted, just as
if it would have been the case if the feature would really be declared
as derived feature.

Bug: 508665
Change-Id: I685e5ab6ef29c3c452845e4f9deae087a9b3b40b
Signed-off-by: Philip Langer <planger@eclipsesource.com>