[508855] Support starting comparisons from containers
When a comparison is started from a container, such as a project or
folder, the EMFModelProvider will now check whether model resources are
contained by the selected container and, if it does contain model
resources, provide resource mappings for them. This enables to start
comparisons from the projects and folders, just as if one would have
selected multiple model files to start the comparison.

To avoid impacting non-model comparisons at all, this functionality is
disabled by default and needs to be enabled with a preference
explicitly.

Bug: 508855
Change-Id: I1007de40dbd18c01482629149f9962da97d82b85
Signed-off-by: Philip Langer <planger@eclipsesource.com>