Merge algorithm refactoring
In order to address several bugs we faced on merging, a refactoring of
the merge algorithm was made.
The responsibility of IMerger is reduced to just merging one diff and
telling which diffs are required to be merged prior to a given diff and
which ones need to be merged in the same 'transaction'.
It is now the responsibility of BatchMerger to merge all the necessary
diffs in the right order when asked to merge one or several diffs.
A new class ComputeDiffsToMerge encapsulates the logic of computing the
diffs to merge and in which order.
The Graph is no longer used to describe the dependencies between diffs,
hence BreadthFirstIterator is no longer used for merging.
Refined diffs are considered as 'to be merged in the same transaction'
(logically resulting merges).
Required diffs must be merged before (merge dependencies).
Equivalent and implied diffs are dealt with like before.
- Diffs have a new state MERGING to distinguish between an in-progress
merge (MERGING) and an already executed merge (MERGED).
- Implied merges need to be set to MERGED recursively
- Diff ordering for implications needs to be checked recursively
- Deterministic refinement order that does not rely on hash values
- Simplified, common method for copyLeftToRight and copyRightToLeft

Includes tests.

Bug: 507177
Bug: 501574
Change-Id: Iad8bad7a4326d4fb297662264b4ad3737f508136
Also-by: Martin Fleck <mfleck@eclipsesource.com>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>