Avoid redundant insertion points
The isVisibleInMergeViewer method now really only checks if a diff is
visible.

Previously the method also checked if the given diff was the
prime-refining diff of another diff and then also returned true. With
the latest StructureMergeViewer and ContentMergeViewer refactorings this
is no longer necessary and leads to duplicated insertion points within
the TreeContentMergeViewers. 

Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I3ecd25abfa7a21b7e8c1e1ee986e36246175dc6e