[499986] Correctly store engine preferences and fix setting to defaults
There was a bug in obtaining the disabled match engines preventing the
selection to be stored.

This change also fixes the behavior of setting the engine page to
defaults. If there is a default configuration of an engine, e.g. using
the plugin_customization, the "Set to defaults" button didn't take this
configuration into account, but always sets the highest ranked engine.

Moreover, this preference page has been refactored to make it more
readable.

Bug: 499986
Change-Id: I37a41f0e872c1fc69b946507e9d611e483b92a2b
Also-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>