Fix verification in logging preference page
Previously, it was almost impossible to modify "Maximum file size" and
"Maximum number of backup file". The verification was blocking any
attempt to replace the value. With this fix we now check if these fields
are valid and use the error message mechanism in Eclipse preference page
to warn the user if these fields are invalid.

Change-Id: I56ea1011ad0914e5a8144ed54cd909053526559c
Signed-off-by: Simon Delisle <simon.delisle@ericsson.com>