[516524] Use more efficient feature look-up in ReferenceUtil
ReferenceUtil could make use of an EClass.getFeatureID(feature) which
returns -1 if the feature doesn't belong to the class. Only in that case
do we need to try to look up the feature by name. Once we've computed
the feature ID, we can use it directly to efficiently access the
generated reflective methods.

Bug: 516524
Change-Id: I6ff031ebc02dd10bf3b0d88c00f03f15c25cfb61
Signed-off-by: Philip Langer <planger@eclipsesource.com>