[516572] Change models' superclass to MinimalEObjectImpl$Container
As we heavily rely on containment structures in the comparison model and
the tree node model (navigating to containers, etc.), it would be better
to extend MinimalEObjectImpl$Container rather than using
MinimalEObjectImpl as it is now. MinimalEObjectImpl$Container has a
dedicated field to store the container rather than using the storage
array also for the containment feature.

Bug: 516572
Change-Id: I4ea7a18963d4f0785f8ef075394388689c67c107
Signed-off-by: Philip Langer <planger@eclipsesource.com>