Refactorings to improve performance of EqualityHelper
EqualityHelper.matchValues is heavily used during comparison so it
should be optimal.
Therefore, this change factors out the null cases so that no null checks
are needed anywhere else; such a check are cheap, compared to instanceof
check so better it be checked early, just as object1 == object2 is done
early.

Change-Id: If62ab5b817307bd9077b324362c02323d044cdd8
Signed-off-by: Philip Langer <planger@eclipsesource.com>