[Performance] Loop invariants and other redundant method call and casts
This change introduces small improvements, which are mostly only minors
but which have a remarkable impact on performance in large comparisons,
as the changed parts are hot spots. This includes method call inside a
loop that always returns the same value as well as repeated calls to the
same method should be avoided.

Change-Id: I2116c1f1f7fed4b946aafa740c4c0c96f7f511b8
Signed-off-by: Philip Langer <planger@eclipsesource.com>