[516645] Fix memory leak in EMFCompareContentMergeViewer
With this change, we do the following on dispose:
menuService.releaseContributions(toolBarManager);
menuService.populateContributionManager(toolBarManager, "nothing"); 
menuService.releaseContributions(toolBarManager);

The first call clears the contributions of the toolbar manager. The
second populates again but with an empty toolbar manager and a location
that presumably has no contributions, resulting in a still empty toolbar
manager, but the important point is that it it reconciles with the
current state, ensuring that all the old items are cleared.

I'm not sure if this situation represents a bug in the IMenuService
implementation or if EMF Compare is abusing it. Anyway, this change
fixes the leak on EMF Compare side.

Bug: 516645
Change-Id: I8f1b07781656e828ac8b6eca0f3f7a36be7cb0ae
Signed-off-by: Philip Langer <planger@eclipsesource.com>