[518755] Mark resources as conflicting on errors during the comparison
If there is a diagnostic error in the comparison model during a git
merge, we should mark all involved files in the resource mapping as
conflicting and don't try to actually merge them, because otherwise we
risk of breaking the model. Therefore, we rather not touch those files
and force manual merge.

Note that we still don't (and don't want to) stop if there are
diagnostic errors in one or more of the compared models (in contrast to
the comparison). Errors in the models most likely don't stop us from
doing a decent merge.

This change also improves the way diagnostics are set up, to avoid the
OK message even though later steps in the comparison may add warnings or
errors to the parent diagnostics for comparisons based on the
IdentifierEObjectMatcher. Other matchers don't create a diagnostic at
all. In this case, the first other step will set a new one instead of
adding their diagnostics to the existing one.

Bug: 518755
Change-Id: I209505dec67daed825f4ba63f40b1f10a79ae36c
Signed-off-by: Philip Langer <planger@eclipsesource.com>